refactor: optimize css
