+++ src/components/Header/Header.css
@@ -1,14 +1,7 @@
 .header {
   box-shadow: rgba(149, 157, 165, 0.2) 0px 8px 24px;
 }
-.container {
-  max-width: 1110px;
-  margin: 0 auto;
-  padding: 0 10px;
-}
 .header__wrapper {
-  display: flex;
-  align-items: center;
   justify-content: space-between;
 }
 .menu {
