+++ src/index.tsx
@@ -4,7 +4,7 @@ import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
 import { createBrowserRouter, RouterProvider } from 'react-router-dom';
-import NotFound from './NotFound';
+import NotFound from './pages/NotFound/NotFound';
 import About from './pages/About';
 import Home from './pages/Home';
 const router = createBrowserRouter([
