+++ src/pages/NotFound/NotFound.css
@@ -1,0 +1,4 @@
+.not-found {
+  flex-direction: column;
+  gap: 20px;
+}
